Remove LSIF HTTP provider.
