+++ shared/lsif/providers.ts
@@ -1,27 +1,36 @@
 import * as sourcegraph from 'sourcegraph'
+import * as lsp from 'vscode-languageserver-protocol'
+import { Providers, noopProviders } from '../providers'
+import { queryGraphQL as sgQueryGraphQL, QueryGraphQLFn } from '../util/graphql'
+import { asyncGeneratorFromPromise, concat } from '../util/ix'
+import { parseGitURI } from '../util/uri'
+import { LocationConnectionNode, nodeToLocation } from './conversion'
 import { Logger } from '../logging'
-import { noopProviders, Providers } from '../providers'
-import { productVersion } from '../util/api'
-import { compareVersion } from '../util/versions'
-import { createProviders as createGraphQLProviders } from './graphql'
-import { createProviders as createHTTPProviders } from './http'
 
 /**
- * The date that the LSIF GraphQL API resolvers became available.
+ * The maximum number of chained GraphQL requests to make for a single
+ * requests query. The page count for a result set should generally be
+ * relatively low unless it's a VERY popular library and LSIF data is
+ * ubiquitous (which is our goal).
+ */
+export const MAX_REFERENCE_PAGE_REQUESTS = 10
+
+/** The response envelope for all LSIF queries. */
+export interface GenericLSIFResponse<R> {
+    repository: {
+        commit: {
+            blob: {
+                lsif: R
+            }
+        }
+    }
+}
+
+/**
+ * Creates providers powered by LSIF-based code intelligence. This particular
+ * set of providers will use the GraphQL API.
  *
- * Specifically, ensure that the commit 34e6a67ecca30afb4a5d8d200fc88a724d3c4ac5
- * exists, as there is a bad performance issue prior to that when a force push
- * removes commits from the codehost for which we have LSIF data.
- */
-const GRAPHQL_API_MINIMUM_DATE = '2020-01-08'
-
-/**
- * The version that the LSIF GraphQL API resolvers became available.
- */
-const GRAPHQL_API_MINIMUM_VERSION = '3.12.0'
-
-/**
- * Creates providers powered by LSIF-based code intelligence.
+ * @param logger The logger instance.
  */
 export function createProviders(logger: Logger): Providers {
     const enabled = !!sourcegraph.configuration.get().get('codeIntel.lsif')
@@ -30,55 +39,289 @@ export function createProviders(logger: Logger): Providers {
         return noopProviders
     }
 
-    const provider = selectProvider(logger)
+    logger.log('LSIF is enabled')
+    return createGraphQLProviders()
+}
 
-    async function* definition(
-        doc: sourcegraph.TextDocument,
-        pos: sourcegraph.Position
-    ): AsyncGenerator<sourcegraph.Definition, void, undefined> {
-        yield* (await provider).definition(doc, pos)
-    }
-
-    async function* references(
-        doc: sourcegraph.TextDocument,
-        pos: sourcegraph.Position,
-        ctx: sourcegraph.ReferenceContext
-    ): AsyncGenerator<sourcegraph.Location[] | null, void, undefined> {
-        yield* (await provider).references(doc, pos, ctx)
-    }
-
-    async function* hover(
-        doc: sourcegraph.TextDocument,
-        pos: sourcegraph.Position
-    ): AsyncGenerator<sourcegraph.Hover | null, void, undefined> {
-        yield* (await provider).hover(doc, pos)
-    }
-
+/**
+ * Creates providers powered by LSIF-based code intelligence. This particular
+ * set of providers will use the GraphQL API.
+ *
+ * @param queryGraphQL The function used to query the GraphQL API.
+ */
+export function createGraphQLProviders(
+    queryGraphQL: QueryGraphQLFn<any> = sgQueryGraphQL
+): Providers {
     return {
-        definition,
-        references,
-        hover,
+        definition: asyncGeneratorFromPromise(definition(queryGraphQL)),
+        references: references(queryGraphQL),
+        hover: asyncGeneratorFromPromise(hover(queryGraphQL)),
+    }
+}
+
+export interface DefinitionResponse {
+    definitions: { nodes: LocationConnectionNode[] }
+}
+
+/** Retrieve a definition for the current hover position. */
+function definition(
+    queryGraphQL: QueryGraphQLFn<GenericLSIFResponse<DefinitionResponse | null>>
+): (
+    doc: sourcegraph.TextDocument,
+    position: sourcegraph.Position
+) => Promise<sourcegraph.Definition> {
+    return async (
+        doc: sourcegraph.TextDocument,
+        position: sourcegraph.Position
+    ): Promise<sourcegraph.Definition> => {
+        const query = `
+            query Definitions($repository: String!, $commit: String!, $path: String!, $line: Int!, $character: Int!) {
+                repository(name: $repository) {
+                    commit(rev: $commit) {
+                        blob(path: $path) {
+                            lsif {
+                                definitions(line: $line, character: $character) {
+                                    nodes {
+                                        resource {
+                                            path
+                                            repository {
+                                                name
+                                            }
+                                            commit {
+                                                oid
+                                            }
+                                        }
+                                        range {
+                                            start {
+                                                line
+                                                character
+                                            }
+                                            end {
+                                                line
+                                                character
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        `
+
+        const lsifObj: DefinitionResponse | null = await queryLSIF(
+            { doc, position, query },
+            queryGraphQL
+        )
+        return lsifObj?.definitions.nodes.map(nodeToLocation) || null
+    }
+}
+
+export interface ReferencesResponse {
+    references: {
+        nodes: LocationConnectionNode[]
+        pageInfo: {
+            endCursor?: string
+        }
+    }
+}
+
+/** Retrieve references for the current hover position. */
+function references(
+    queryGraphQL: QueryGraphQLFn<GenericLSIFResponse<ReferencesResponse | null>>
+): (
+    doc: sourcegraph.TextDocument,
+    position: sourcegraph.Position
+) => AsyncGenerator<sourcegraph.Location[] | null, void, undefined> {
+    // eslint-disable-next-line @typescript-eslint/require-await
+    return async function*(
+        doc: sourcegraph.TextDocument,
+        position: sourcegraph.Position
+    ): AsyncGenerator<sourcegraph.Location[] | null, void, undefined> {
+        const query = `
+            query References($repository: String!, $commit: String!, $path: String!, $line: Int!, $character: Int!, $after: String) {
+                repository(name: $repository) {
+                    commit(rev: $commit) {
+                        blob(path: $path) {
+                            lsif {
+                                references(line: $line, character: $character, after: $after) {
+                                    nodes {
+                                        resource {
+                                            path
+                                            repository {
+                                                name
+                                            }
+                                            commit {
+                                                oid
+                                            }
+                                        }
+                                        range {
+                                            start {
+                                                line
+                                                character
+                                            }
+                                            end {
+                                                line
+                                                character
+                                            }
+                                        }
+                                    }
+                                    pageInfo {
+                                        endCursor
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        `
+
+        const queryPage = async function*(
+            requestsRemaining: number,
+            after?: string
+        ): AsyncGenerator<sourcegraph.Location[] | null, void, undefined> {
+            if (requestsRemaining === 0) {
+                return
+            }
+
+            // Make the request for the page starting at the after cursor
+            const lsifObj: ReferencesResponse | null = await queryLSIF(
+                {
+                    doc,
+                    position,
+                    query,
+                    after,
+                },
+                queryGraphQL
+            )
+            if (!lsifObj) {
+                return
+            }
+
+            const {
+                references: {
+                    nodes,
+                    pageInfo: { endCursor },
+                },
+            } = lsifObj
+
+            // Yield this page's set of results
+            yield nodes.map(nodeToLocation)
+
+            if (endCursor) {
+                // Recursively yield the remaining pages
+                yield* queryPage(requestsRemaining - 1, endCursor)
+            }
+        }
+
+        yield* concat(queryPage(MAX_REFERENCE_PAGE_REQUESTS))
+    }
+}
+
+export interface HoverResponse {
+    hover?: { markdown: { text: string }; range: sourcegraph.Range }
+}
+
+/** Retrieve hover text for the current hover position. */
+function hover(
+    queryGraphQL: QueryGraphQLFn<GenericLSIFResponse<HoverResponse | null>>
+): (
+    doc: sourcegraph.TextDocument,
+    position: sourcegraph.Position
+) => Promise<sourcegraph.Hover | null> {
+    return async (
+        doc: sourcegraph.TextDocument,
+        position: sourcegraph.Position
+    ): Promise<sourcegraph.Hover | null> => {
+        const query = `
+            query Hover($repository: String!, $commit: String!, $path: String!, $line: Int!, $character: Int!) {
+                repository(name: $repository) {
+                    commit(rev: $commit) {
+                        blob(path: $path) {
+                            lsif {
+                                hover(line: $line, character: $character) {
+                                    markdown {
+                                        text
+                                    }
+                                    range {
+                                        start {
+                                            line
+                                            character
+                                        }
+                                        end {
+                                            line
+                                            character
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        `
+
+        const lsifObj: HoverResponse | null = await queryLSIF(
+            {
+                doc,
+                position,
+                query,
+            },
+            queryGraphQL
+        )
+
+        if (!lsifObj || !lsifObj.hover) {
+            return null
+        }
+
+        return {
+            contents: {
+                value: lsifObj.hover.markdown.text,
+                kind: sourcegraph.MarkupKind.Markdown,
+            },
+            range: lsifObj.hover.range,
+        }
     }
 }
 
 /**
- * Return the GraphQL LSIF providers if the Sourcegraph instance supports it.
- * Otherwise, use the HTTP API providers.
+ * Perform an LSIF request to the GraphQL API.
+ *
+ * @param args Parameter bag.
+ * @param queryGraphQL The function used to query the GraphQL API.
  */
-async function selectProvider(logger: Logger): Promise<Providers> {
-    const supportsGraphQL = compareVersion({
-        productVersion: await productVersion(),
-        minimumVersion: GRAPHQL_API_MINIMUM_VERSION,
-        minimumDate: GRAPHQL_API_MINIMUM_DATE,
-    })
-
-    if (supportsGraphQL) {
-        logger.log('Sourcegraph instance supports LSIF GraphQL API')
-        return createGraphQLProviders()
+async function queryLSIF<
+    P extends {
+        doc: sourcegraph.TextDocument
+        position: lsp.Position
+        query: string
+    },
+    R
+>(
+    {
+        /** The current text document. */
+        doc,
+        /** The current hover position. */
+        position,
+        /** The GraphQL request query. */
+        query,
+        /** Additional query parameters. */
+        ...rest
+    }: P,
+    queryGraphQL: QueryGraphQLFn<GenericLSIFResponse<R>>
+): Promise<R | null> {
+    const { repo, commit, path } = parseGitURI(new URL(doc.uri))
+    const queryArgs = {
+        repository: repo,
+        commit,
+        path,
+        line: position.line,
+        character: position.character,
+        ...rest,
     }
 
-    logger.log(
-        'Sourcegraph instance does not support LSIF GraphQL API, falling back to HTTP API'
-    )
-    return createHTTPProviders()
+    const data = await queryGraphQL(query, queryArgs)
+    return data.repository.commit.blob.lsif
 }
